refactoring map.jsx
