+++ public/data/Locations.json
@@ -20,11 +20,6 @@
         "lng": 13.77678
     },
     {
-        "name": "Conversano",
-        "lat": 40.96836,
-        "lng": 17.11329
-    },
-    {
         "name": "Poznań",
         "lat": 52.40692,
         "lng": 16.92993
